Define a concern for commentable models dependent of components
