+++ decidim-accountability/app/models/decidim/accountability/result.rb
@@ -12,7 +12,7 @@ module Decidim
       include Decidim::ScopableResource
       include Decidim::HasCategory
       include Decidim::HasReference
-      include Decidim::Comments::Commentable
+      include Decidim::Comments::CommentableWithComponent
       include Decidim::Traceable
       include Decidim::Loggable
       include Decidim::DataPortability
@@ -66,16 +66,6 @@ module Decidim
         save!
       end
 
-      # Public: Overrides the `commentable?` Commentable concern method.
-      def commentable?
-        component.settings.comments_enabled?
-      end
-
-      # Public: Overrides the `accepts_new_comments?` Commentable concern method.
-      def accepts_new_comments?
-        commentable? && !component.current_settings.comments_blocked
-      end
-
       # Public: Overrides the `comments_have_alignment?` Commentable concern method.
       def comments_have_alignment?
         true
@@ -84,11 +74,6 @@ module Decidim
       # Public: Overrides the `comments_have_votes?` Commentable concern method.
       def comments_have_votes?
         true
-      end
-
-      # Public: Whether the object can have new comments or not.
-      def user_allowed_to_comment?(user)
-        can_participate_in_space?(user)
       end
 
       ransacker :id_string do
